refactor: 입력 형식 관련 예외의 부모 클래스 구현
