+++ src/main/java/woowacourse/auth/exception/format/FormatException.java
@@ -1,0 +1,7 @@
+package woowacourse.auth.exception.format;
+
+public class FormatException extends RuntimeException {
+    public FormatException(String message) {
+        super(message);
+    }
+}
