+++ src/main/java/woowacourse/auth/exception/format/InvalidAddressFormatException.java
@@ -1,6 +1,6 @@
-package woowacourse.auth.exception;
+package woowacourse.auth.exception.format;
 
-public class InvalidAddressFormatException extends RuntimeException {
+public class InvalidAddressFormatException extends FormatException {
     public InvalidAddressFormatException() {
         super("올바르지 않은 주소 형식입니다.");
     }
