+++ src/main/java/woowacourse/auth/exception/format/InvalidNameFormatException.java
@@ -1,6 +1,6 @@
-package woowacourse.auth.exception;
+package woowacourse.auth.exception.format;
 
-public class InvalidNameFormatException extends RuntimeException {
+public class InvalidNameFormatException extends FormatException {
     public InvalidNameFormatException() {
         super("올바르지 않은 이름 형식입니다.");
     }
