+++ src/main/java/woowacourse/auth/exception/format/InvalidPasswordFormatException.java
@@ -1,7 +1,6 @@
-package woowacourse.auth.exception;
+package woowacourse.auth.exception.format;
 
-public class InvalidPasswordFormatException extends RuntimeException {
-
+public class InvalidPasswordFormatException extends FormatException {
     public InvalidPasswordFormatException() {
         super("올바르지 않은 패스워드 입력 형식입니다.");
     }
