+++ src/main/java/woowacourse/auth/exception/format/InvalidUrlFormatException.java
@@ -1,6 +1,6 @@
-package woowacourse.auth.exception;
+package woowacourse.auth.exception.format;
 
-public class InvalidUrlFormatException extends RuntimeException {
+public class InvalidUrlFormatException extends FormatException {
     public InvalidUrlFormatException() {
         super("올바르지 않은 URL 형식입니다.");
     }
