+++ src/main/java/woowacourse/auth/exception/format/InvalidZoneCodeFormatException.java
@@ -1,6 +1,6 @@
-package woowacourse.auth.exception;
+package woowacourse.auth.exception.format;
 
-public class InvalidZoneCodeFormatException extends RuntimeException {
+public class InvalidZoneCodeFormatException extends FormatException {
     public InvalidZoneCodeFormatException() {
         super("올바르지 않은 우편번호 형식입니다.");
     }
