+++ src/test/java/woowacourse/auth/domain/user/AddressTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import woowacourse.auth.exception.InvalidAddressFormatException;
+import woowacourse.auth.exception.format.InvalidAddressFormatException;
 
 class AddressTest {
     @DisplayName("주소 문자열을 전달받아 생성된다.")
