+++ src/test/java/woowacourse/auth/domain/user/EmailTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import woowacourse.auth.exception.InvalidEmailFormatException;
+import woowacourse.auth.exception.format.InvalidEmailFormatException;
 
 class EmailTest {
