+++ src/test/java/woowacourse/auth/domain/user/NameTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import woowacourse.auth.exception.InvalidNameFormatException;
+import woowacourse.auth.exception.format.InvalidNameFormatException;
 
 class NameTest {
