+++ src/test/java/woowacourse/auth/domain/user/ProfileImageUrlTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import woowacourse.auth.exception.InvalidUrlFormatException;
+import woowacourse.auth.exception.format.InvalidUrlFormatException;
 
 class ProfileImageUrlTest {
